Less discarding of test failure throwables
... and more propagating thereof.

On one end, it seems a shame to discard information that we have.

On the other end, receiver of the `sbt.testing.Event` - in my case, [Gradle plugin for multi-backend Scala](https://github.com/dubinsky/scalajs-gradle) - has to provide a throwable to Gradle, which requires one to accompany a test failure, and if it is not propagated by the test framework, a fake throwable has to be concocted (ugh).

This change makes the ends meet ;)